requirements.txt
numpy
Pillow
scikit-image
matplotlib
pandas
openpyxl

